requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.12
pymongo==4.5.1
celery==5.3.1
redis==5.4.2
python-dotenv==1.0.0
playwright==1.36.0
beautifulsoup4==4.12.2
fake-useragent==1.1.3
requests==2.31.0
python-multipart==0.0.6
dnspython==2.3.0
orjson==3.9.1
aiohttp==3.8.4
python-dateutil==2.8.2
tqdm==4.65.0
